+++ src/cart/cart.module.ts
@@ -5,9 +5,9 @@ import { OrderModule } from '../order/order.module';
 import { CartController } from './cart.controller';
 import { CartService } from './services';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Carts } from './entity/Carts';
-import { CartItems } from './entity/CartItems';
-import { Orders } from '../order/entity/Order';
+import { Carts } from '../entities/Carts';
+import { CartItems } from '../entities/CartItems';
+import { Orders } from '../entities/Order';
 
 @Module({
   imports: [OrderModule, TypeOrmModule.forFeature([Carts, CartItems, Orders])],
